chore: Set up project folder structure
