+++ src/main.js
@@ -1,16 +1,29 @@
-export class Car {
-  name;
-  position = 0;
+import readline from 'readline';
 
-  constructor(name) {
-    if (!name) {
-      throw new Error('car name is required');
+function readLineAsync(query) {
+  return new Promise((resolve, reject) => {
+    if (arguments.length !== 1) {
+      reject(new Error('arguments must be 1'));
     }
 
-    this.name = name;
-  }
+    if (typeof query !== 'string') {
+      reject(new Error('query must be string'));
+    }
 
-  move() {
-    this.position += 1;
-  }
+    const rl = readline.createInterface({
+      input: process.stdin,
+      output: process.stdout,
+    });
+
+    rl.question(query, (input) => {
+      rl.close();
+      resolve(input);
+    });
+  });
+}
+
+// 입출력 예시
+async function play() {
+  const name = await readLineAsync('자동차 이름을 입력하세요 > ');
+  console.log(`자동차 이름은 ${name}입니다.`);
 }
